Fix flaky unit test to parse instanceID
Summary: Util method to parse instance id returned int rather than uint32, and in error case value return was -1. It was leading to flaky unit test.

Reviewers: zhixin, apoorvaj, mu, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2093993

+++ hostmgr/factory/task/task_builder_test.go
@@ -704,7 +704,8 @@ func (suite *BuilderTestSuite) TestPopulateLabels() {
 
 	for _, test := range tt {
 		mesosTask := &mesos.TaskInfo{}
-		builder.populateLabels(mesosTask, test.pelotonTaskLabels, jobID, instanceID)
+		builder.populateLabels(
+			mesosTask, test.pelotonTaskLabels, jobID, uint32(instanceID))
 		suite.Equal(test.mesosTaskLabels, mesosTask.Labels)
 	}
 }
